+++ source/src/main/java/org/cerberus/crud/factory/IFactoryScheduleEntry.java
@@ -1,6 +1,5 @@
 package org.cerberus.crud.factory;
 
-import org.cerberus.crud.entity.CampaignParameter;
 
 /* Cerberus Copyright (C) 2013 - 2017 cerberustesting
 DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
@@ -19,18 +18,29 @@ GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License
 along with Cerberus.  If not, see <http://www.gnu.org/licenses/>.
-*/
- 
+ */
 /**
  *
  * @author cdelage
  */
 import org.cerberus.crud.entity.ScheduleEntry;
 import java.sql.Timestamp;
-import org.cerberus.crud.factory.impl.FactoryScheduleEntry;
-import org.springframework.beans.factory.annotation.Autowired;
 
 public interface IFactoryScheduleEntry {
-    
-    ScheduleEntry create(Integer ID, String type, String name, String cronDefinition, Timestamp lastExecution, String active, String UsrCreated, Timestamp DateCreated, String UsrModif, Timestamp DateModif);
+
+    /**
+     *
+     * @param ID
+     * @param type
+     * @param name
+     * @param cronDefinition
+     * @param lastExecution
+     * @param active
+     * @param UsrCreated
+     * @param DateCreated
+     * @param UsrModif
+     * @param DateModif
+     * @return
+     */
+    ScheduleEntry create(long ID, String type, String name, String cronDefinition, Timestamp lastExecution, String active, String UsrCreated, Timestamp DateCreated, String UsrModif, Timestamp DateModif);
 }
